requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.4
jinja2==3.1.4
pydantic==2.8.2
python-dotenv==1.0.1

langchain==0.1.20
langchain-openai==0.0.8
